Fixes a crash and fixes warnings
Fixes a warning where an 8 byte entity_t is being truncated to a 4 byte unsigned long.  Changed to an unsigned long long which is 64bit wide on all systems.

Also fixed a crash where a forced unsigned 64bit integer is being used for the vid.width buffer.  Not a good idea.

Also fixed a crash when exiting because you should not quit SDL3 'after' assets and resources have been free'd..  this should be done 'before',